back-end/requirements.txt
flask
pandas
sqlalchemy
psycopg2-binary
requests
python-dotenv
pyjwt[crypto]